requirements.txt
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cripto==1.0a0
cryptography==3.4.6
Django==3.1.7
django-appconf==1.0.4
django-cryptography==1.0
django-firebase-authentication==1.0.9
djangorestframework==3.12.4
firebase-admin==4.5.3
firebase-orm==0.1.29
gcloud==0.18.3
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
httplib2==0.19.0
idna==2.10
jws==0.1.3
msgpack==1.0.2
oauth2client==4.1.3
packaging==20.9
Pillow==8.1.2
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase==3.0.27
Pyrebase4==4.4.3
python-firebase==1.2
python-jwt==2.0.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
sseclient==0.0.27
uritemplate==3.0.1
urllib3==1.26.4
